Spike out automatic standard route generation
