+++ augur/metrics/insight.py
@@ -6,8 +6,7 @@ import sqlalchemy as s
 import pandas as pd
 from augur.util import register_metric
 
-
-@register_metric()
+@register_metric(type="repo_group_only")
 def top_insights(self, repo_group_id, num_repos=6):
     """
     Timeseries of pull request acceptance rate (expressed as the ratio of pull requests merged on a date to the count of pull requests opened on a date)
