+++ augur/metrics/metrics.py
@@ -11,16 +11,14 @@ class Metrics():
         self.database = app.database
         self.spdx_db = app.spdx_db
 
-        models = [] #TODO: standardize this
+        self.models = [] #TODO: standardize this
         for filename in glob.iglob("augur/metrics/**"):
             file_id = get_file_id(filename)
             if not file_id.startswith('__') and filename.endswith('.py') and file_id != "metrics":
-                models.append(file_id)
+                self.models.append(file_id)
 
-        for model in models:
+        for model in self.models:
             importlib.import_module(f"augur.metrics.{model}")
-
-        for model in models:
             add_metrics(self, f"augur.metrics.{model}")
 
 def get_file_id(path):
@@ -30,9 +28,8 @@ def add_metrics(metrics, module_name):
     # find all unbound endpoint functions objects 
     # (ones that have metadata) defined the given module_name 
     # and bind them to the metrics class
-    # Derek are you proud of me
     for name, obj in inspect.getmembers(sys.modules[module_name]):
         if inspect.isfunction(obj) == True:
-            if hasattr(obj, 'metadata') == True:
+            if hasattr(obj, 'is_metric') == True:
                 setattr(metrics, name, types.MethodType(obj, metrics))
