--- augur/routes/metrics/issue.py
@@ -1,39 +1,0 @@
-#SPDX-License-Identifier: MIT
-
-def create_routes(server):
-
-    metrics = server.augur_app.metrics
-
-    server.add_standard_metric(metrics.issues_new, 'issues-new')
-
-    server.add_standard_metric(metrics.issues_active, 'issues-active')
-
-    server.add_standard_metric(metrics.issues_closed, 'issues-closed')
-
-    server.add_standard_metric(metrics.issue_duration, 'issue-duration')
-
-    server.add_standard_metric(metrics.issue_participants, 'issue-participants')
-
-    server.add_standard_metric(metrics.issue_backlog, 'issue-backlog')
-
-    server.add_standard_metric(metrics.issue_throughput, 'issue-throughput')
-
-    server.add_standard_metric(metrics.issues_first_time_opened, 'issues-first-time-opened')
-
-    server.add_standard_metric(metrics.issues_first_time_closed, 'issues-first-time-closed')
-
-    server.add_standard_metric(metrics.open_issues_count, 'open-issues-count')
-
-    server.add_standard_metric(metrics.closed_issues_count, 'closed-issues-count')
-
-    server.add_standard_metric(metrics.issues_open_age, 'issues-open-age')
-
-    server.add_standard_metric(metrics.issues_closed_resolution_duration, 'issues-closed-resolution-duration')
-
-    server.add_standard_metric(metrics.issues_maintainer_response_duration, 'issues-maintainer-response-duration')
-
-    server.add_standard_metric(metrics.average_issue_resolution_time, 'average-issue-resolution-time')
-
-    server.add_standard_metric(metrics.issue_comments_mean, 'issue-comments-mean')
-
-    server.add_standard_metric(metrics.issue_comments_mean_std, 'issue-comments-mean-std')
