--- augur/routes/metrics/message.py
@@ -1,6 +1,0 @@
-
-def create_routes(server):
-
-    metrics = server.augur_app.metrics
-
-
