--- augur/routes/metrics/platform.py
@@ -1,4 +1,0 @@
-
-def create_routes(server):
-    metrics = server.augur_app.metrics
-
