+++ augur/server.py
@@ -3,12 +3,18 @@
 Creates a WSGI server that serves the Augur REST API
 """
 
+import glob
+import sys
+import inspect
+import types
 import json
 import os
 import base64
+
 from flask import Flask, request, Response, redirect
 from flask_cors import CORS
 import pandas as pd
+
 import augur
 from augur.util import logger
 from augur.routes import create_routes
